Enforce `dpdk` `mbuf` functions inlining
There are shot `mbuf` manipulation functions wrapped inside
`packet.c` source file. As it was discovered some of them like
`packet_data_len` consumes relatively high CPU clock count while pipeline
processing.
