+++ dataplane/dataplane.c
@@ -25,6 +25,9 @@
 #include "dataplane/device.h"
 #include "dataplane/worker.h"
 
+#include "lib/dataplane/packet/data.h"
+#include "lib/dataplane/packet/packet.h"
+
 #include <unistd.h>
 
 #include "sys/mman.h"
