+++ dataplane/worker.c
@@ -42,6 +42,9 @@
 #include "dataplane/device.h"
 
 #include "lib/dataplane/config/zone.h"
+#include "lib/dataplane/module/packet_front.h"
+#include "lib/dataplane/packet/data.h"
+#include "lib/dataplane/packet/packet.h"
 #include "lib/dataplane/pipeline/pipeline.h"
 #include "lib/dataplane/time/clock.h"
