+++ lib/fwstate/sync.c
@@ -8,6 +8,7 @@
 
 #include "common/network.h"
 
+#include "lib/dataplane/packet/data.h"
 #include "lib/dataplane/packet/packet.h"
 #include "lib/dataplane/worker/worker.h"
 #include "sync.h"
