+++ lib/utils/go/packet.go
@@ -5,6 +5,7 @@ package dataplane
 //#cgo LDFLAGS: -L../../../build/lib/dataplane/packet -lpacket
 //
 //#include "lib/dataplane/packet/packet.h"
+//#include "lib/dataplane/module/packet_front.h"
 //#include "lib/utils/packet.h"
 import "C"
 import (
