+++ mock/worker.c
@@ -5,6 +5,7 @@
 #include "dataplane/config/zone.h"
 #include "dataplane/pipeline/pipeline.h"
 #include "dataplane/time/clock.h"
+#include "lib/dataplane/module/packet_front.h"
 #include "packet.h"
 
 ////////////////////////////////////////////////////////////////////////////////
