+++ modules/decap/dataplane/dataplane.c
@@ -3,6 +3,9 @@
 #include "config.h"
 
 #include "dataplane/module/module.h"
+#include "lib/dataplane/module/packet_front.h"
+#include "lib/dataplane/packet/data.h"
+#include "lib/dataplane/packet/packet.h"
 
 #include "dataplane/packet/decap.h"
