+++ modules/fwstate/dataplane/dataplane.c
@@ -11,6 +11,9 @@
 #include "dataplane/module/module.h"
 #include "fwstate/layermap.h"
 #include "fwstate/types.h"
+#include "lib/dataplane/module/packet_front.h"
+#include "lib/dataplane/packet/data.h"
+#include "lib/dataplane/packet/packet.h"
 #include "lib/dataplane/time/clock.h"
 #include "logging/log.h"
