+++ modules/nat64/dataplane/nat64dp.c
@@ -47,8 +47,10 @@
 /* Project headers */
 #include "common.h"
 #include "dataplane/module/module.h"
+#include "dataplane/packet/data.h"
 #include "dataplane/packet/icmp.h"
 #include "dataplane/packet/packet.h"
+#include "lib/dataplane/module/packet_front.h"
 #include "modules/nat64/dataplane/nat64dp.h"
 
 /**
